backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
email-validator==2.1.0
python-multipart==0.0.6
python-dateutil==2.8.2
python-dotenv==1.0.0



